edit few elements in advanced settings interface #29935
