+++ projects/back-office/src/app/dashboard/pages/advanced-settings/advanced-settings-routing.module.ts
@@ -2,6 +2,7 @@ import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { AdvancedSettingsComponent } from './advanced-settings.component';
 
+/** Routes of Advanced settings page */
 const routes: Routes = [
   {
     path: '',
@@ -9,6 +10,9 @@ const routes: Routes = [
   },
 ];
 
+/**
+ * Advanced settings page routing module.
+ */
 @NgModule({
   imports: [RouterModule.forChild(routes)],
   exports: [RouterModule],
