+++ projects/back-office/src/app/dashboard/pages/advanced-settings/advanced-settings.component.html
@@ -1,44 +1,46 @@
 <div class="page-header">
-  <safe-previous-button path="/settings/users"></safe-previous-button>
   <h1 class="page-title">
     {{ 'components.users.advancedSettings.title' | translate }}
   </h1>
 </div>
 <mat-spinner diameter="45" class="page-loader" *ngIf="loading"></mat-spinner>
+<h2>
+  {{ 'components.users.advancedSettings.userManagement.title' | translate }}
+</h2>
 <form [formGroup]="settingForm" *ngIf="!loading">
-  <h2>
-    {{ 'components.users.advancedSettings.userManagement.title' | translate }}
-  </h2>
-  <div formGroupName="userManagement" class="user-management">
+  <div formGroupName="userManagement" class="form-group">
     <mat-radio-group formControlName="local" class="form-radio-group">
-      <div class="radio-button-tooltip">
-        <mat-radio-button [value]="true">{{
+      <mat-radio-button [value]="true"
+        >{{
           'components.users.advancedSettings.userManagement.local.title'
             | translate
-        }}</mat-radio-button>
-        <mat-icon
-          class="info-icon"
+        }}<safe-icon
+          [inline]="true"
+          icon="info_outline"
+          variant="grey"
+          [size]="18"
           [matTooltip]="
             'components.users.advancedSettings.userManagement.local.tooltip'
               | translate
           "
-          >info</mat-icon
-        >
-      </div>
-      <div class="radio-button-tooltip">
-        <mat-radio-button [value]="false">{{
+        ></safe-icon
+      ></mat-radio-button>
+      <mat-radio-button [value]="false"
+        >{{
           'components.users.advancedSettings.userManagement.external.title'
             | translate
-        }}</mat-radio-button>
-        <mat-icon
-          class="info-icon"
+        }}
+        <safe-icon
+          [inline]="true"
+          icon="info_outline"
+          variant="grey"
+          [size]="18"
           [matTooltip]="
             'components.users.advancedSettings.userManagement.external.tooltip'
               | translate
           "
-          >info</mat-icon
-        >
-      </div>
+        ></safe-icon
+      ></mat-radio-button>
     </mat-radio-group>
     <ng-container
       *ngIf="!$any(settingForm.controls.userManagement).controls.local.value"
@@ -72,19 +74,19 @@
           formControlName="serviceAPI"
         />
       </mat-form-field>
-      <h3>{{ 'components.mapping.title' | translate }}</h3>
-      <safe-mapping [mappingForm]="mappingForm"></safe-mapping>
+      <div>
+        <h3>{{ 'components.mapping.title' | translate }}</h3>
+        <safe-mapping [mappingForm]="mappingForm"></safe-mapping>
+      </div>
     </ng-container>
   </div>
-  <div class="align-end">
-    <safe-button
-      class="save-button"
-      category="secondary"
-      variant="primary"
-      (click)="onSave()"
-      [disabled]="settingForm.invalid || !settingForm.dirty"
-    >
-      {{ 'common.save' | translate }}
-    </safe-button>
-  </div>
+  <safe-button
+    style="align-self: end"
+    category="secondary"
+    variant="primary"
+    (click)="onSave()"
+    [disabled]="settingForm.invalid || !settingForm.dirty"
+  >
+    {{ 'common.save' | translate }}
+  </safe-button>
 </form>
