+++ projects/back-office/src/app/dashboard/pages/advanced-settings/advanced-settings.component.scss
@@ -1,0 +1,12 @@
+.form-group {
+  gap: 8px
+}
+
+.api-input {
+  width: 100%;
+}
+
+form {
+  display: flex;
+  flex-direction: column;
+}
