--- projects/back-office/src/app/dashboard/pages/users/advanced-settings/advanced-settings.component.scss
@@ -1,23 +1,0 @@
-.user-management {
-  display: flex;
-  flex-direction: column;
-  margin-left: 5%;
-}
-
-.radio-button-tooltip {
-  display: flex;
-}
-
-.align-end {
-  align-items: flex-end;
-  display: flex;
-  flex-direction: column;
-}
-
-.api-input {
-  width: 100%;
-}
-
-.form-radio-group {
-  margin-bottom: 1.25em;
-}
