+++ projects/back-office/src/app/graphql/mutations.ts
@@ -14,7 +14,6 @@ import {
   ApiConfiguration,
   PullJob,
   ReferenceData,
-  Setting,
 } from '@safe/builder';
 
 // === EDIT USER ===
@@ -1042,26 +1041,4 @@ export const EDIT_REFERENCE_DATA = gql`
 export interface EditReferenceDataMutationResponse {
   loading: boolean;
   editReferenceData: ReferenceData;
-}
-
-// === EDIT SETTING ===
-export const EDIT_SETTING = gql`
-  mutation editSetting($userManagement: JSON) {
-    editSetting(userManagement: $userManagement) {
-      userManagement {
-        local
-        apiConfiguration {
-          id
-          name
-        }
-        serviceAPI
-        attributesMapping
-      }
-    }
-  }
-`;
-
-export interface EditSettingMutationResponse {
-  loading: boolean;
-  editSetting: Setting;
 }
