+++ projects/safe/src/lib/components/mapping/mapping-forms.ts
@@ -1,4 +1,5 @@
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import get from 'lodash/get';
 import { Mapping } from '../../models/setting.model';
 
 /** Creating a new instance of the FormBuilder class. */
@@ -12,8 +13,8 @@ const formBuilder = new FormBuilder();
  */
 export const createFormGroup = (mapping: Mapping | null): FormGroup =>
   formBuilder.group({
-    field: [mapping && mapping.field ? mapping.field : '', Validators.required],
-    path: [mapping && mapping.path ? mapping.path : '', Validators.required],
-    value: [mapping && mapping.value ? mapping.value : '', Validators.required],
-    text: [mapping && mapping.text ? mapping.text : '', Validators.required],
+    field: [get(mapping, 'field', ''), Validators.required],
+    path: [get(mapping, 'path', ''), Validators.required],
+    value: [get(mapping, 'value', ''), Validators.required],
+    text: [get(mapping, 'text', ''), Validators.required],
   });
