+++ projects/safe/src/lib/components/ui/icon/icon.component.scss
@@ -1,7 +1,3 @@
-:host {
-  display: flex;
-}
-
 .safe-icon-success {
   color: #63d168;
 }
