+++ projects/safe/src/lib/components/ui/icon/icon.component.ts
@@ -1,4 +1,4 @@
-import { Component, Input, OnInit } from '@angular/core';
+import { Component, HostBinding, Input, OnInit } from '@angular/core';
 import { IconVariant } from './icon-variant.enum';
 
 /**
@@ -17,6 +17,10 @@ export class SafeIconComponent implements OnInit {
   @Input() variant: IconVariant | string = IconVariant.DEFAULT;
 
   @Input() size = 24;
+
+  @HostBinding('style.display') get display() {
+    return this.inline ? 'inline-flex' : 'flex';
+  }
 
   /**
    * Formats the size input adding a 'px' suffix
